requirements.txt
streamlit==1.24.0
pandas==1.5.3
altair==4.2.2
python-dotenv==0.21.1
openai==0.27.8
matplotlib==3.7.1
Pillow==9.5.0
reportlab==4.0.4
tweepy==4.14.0
facebook-sdk==3.1.0
seaborn==0.12.2
sqlalchemy==1.4.47